refactor: restructure Sidebar component to use server-side data fetching and separate client-side logic into SidebarClient
